requirements_radiation.txt
numpy>=1.21.0
matplotlib>=3.5.0
RPi.GPIO>=0.7.0
readchar>=3.0.0
opencv-python>=4.5.0